Consolidate form API routes into a single registry-driven file
Replace separate cert.js and 849b.js route files with a shared
index.js that iterates over a FORMS registry. Each entry provides
its Prisma includes, data transform, Zod schema, and PDF metadata.
Route handlers, error responses, and the PDF generation pipeline
are now defined once.
