+++ server/lib/pdf.js
@@ -31,7 +31,7 @@ async function getFormContainer () {
  * Renders a React form component to a full HTML document string.
  * The component should render its form content only (no <html>/<body> wrapper);
  * FormContainer provides the document shell in standalone mode.
- * 
+ *
  * @param {React.Component} FormComponent - The form component to render
  * @param {object} data - Data to pass to the form component
  * @param {object} [options] - Additional options
